Main functions

General proximal/projection functionalities

AbstractProximableFunctions.prox — Method

prox(y::AbstractArray{CT,N}, λ::T, g::AbstractProximableFunction{CT,N}) where {T<:Real,N,CT<:RealOrComplex{T}}

Computes the proximal operator $\mathrm{prox}_{\lambda,g}(\mathbf{y})$.

AbstractProximableFunctions.prox — Method

prox(y::AbstractArray{CT,N}, λ::T, g::AbstractProximableFunction{CT,N}, options::AbstractArgminOptions) where {T<:Real,N,CT<:RealOrComplex{T}}

Computes the proximal operator with specified options.

See prox(y::AbstractArray{CT,N}, λ::T, g::AbstractProximableFunction{CT,N}) where {T<:Real,N,CT<:RealOrComplex{T}}

AbstractProximableFunctions.proj — Method

proj(y::AbstractArray{CT,N}, ε::T, g::AbstractProximableFunction{CT,N}) where {T<:Real,N,CT<:RealOrComplex{T}}

Computes the projection operator $\mathrm{proj}_{\lambda,g}(\mathbf{y})$.

AbstractProximableFunctions.proj — Method

proj(y::AbstractArray{CT,N}, ε::T, g::AbstractProximableFunction{CT,N}, options::AbstractArgminOptions) where {T<:Real,N,CT<:RealOrComplex{T}}

Computes the projection operator with specified options.

See proj(y::AbstractArray{CT,N}, ε::T, g::AbstractProximableFunction{CT,N}) where {T<:Real,N,CT<:RealOrComplex{T}}

Iterative solver utilities

AbstractProximableFunctions.exact_argmin — Function

exact_argmin()

Returns exact optimization options for the computation of proximal/projection operators. It might results in an error if the regularization functional does not implement analytically-defined proximal/projection operators (e.g. TV!).

AbstractProximableFunctions.FISTA_options — Function

FISTA_options(L; Nesterov=true,
                 reset_counter=nothing,
                 niter=nothing,
                 verbose=false,
                 fun_history=false)

Returns FISTA iterative solver options for the general optimization problem:

$\min_{\mathbf{x}}f(\mathbf{x})+g(\mathbf{x})$

where $g$ is a "proximable" function. It can be used to define options for the proximal or projection operators.

The parameter $L$ ideally should be chosen as $L\ge\mathrm{Lip}\nabla f$ (=Lipschitz constant of the gradient) and is problem specific.

Nesterov acceleration is set by Nesterov=true, while reset_counter is the number of iteration after which the Nesterov momentum is reset. The total number of iterations is determined by niter.

For debugging, set verbose=true and/or fun_history=true (the latter allows storing the history of $f(\mathbf{x})$, which can be retrieved by fun_history(options) after minimization).

Important note

When setting the options for proximal or projection operators, follow the recommendations for each specific proximal function on how to choose $L$. The underlying optimization may be based on algebraic reformulations of the optimization problem $\min_{\mathbf{x}}1/2||\mathbf{x}-\mathbf{y}||^2+\lambda{}g(\mathbf{x})$, so the assumption that $f(\mathbf{x})=1/2||\mathbf{x}-\mathbf{y}||^2$ is not generally correct. This note is relevant when computing the proximal operator of WeightedProximalFunction's for which the problem $\min_{\mathbf{x}}1/2||\mathbf{x}-\mathbf{y}||^2+\lambda{}g(A\mathbf{x})$ is transformed to $\min_{\mathbf{p}}1/2||\lambda A^*\mathbf{p}-\mathbf{y}||^2+\lambda g^*(\mathbf{p})$. The Lipschitz constant in this case should be $\lambda^2\rho(A)$. For convenience of use, however, the Lipschitz constant is expected to be just $\rho(A)$.

Implemented proximable functions

LinearAlgebra.norm — Method

LinearAlgebra.norm(T::DataType, N::Number, P::Number)

Return the proximable function $g(\mathbf{x})=||\mathbf{x}||_p$ where $p\in\{1,2,\infty\}$.

Must specify the element type $T$ and array dimension $N$ of the input array.

AbstractProximableFunctions.mixed_norm — Method

mixed_norm(T::DataType, D::Number, N1::Number, N2::Number)

Return the mixed-norm proximable function. For a D+1-dimensional input array $\mathbf{x}_{i_1,\ldots,i_D}{}^{i_{D+1}}$, it represents the functional $g(\mathbf{x})=||(i_1,\ldots,i_D)\mapsto||\mathbf{x}_{i_1,\ldots,i_D}{}^:||_{\ell^{N_1}}||_{\ell^{N_2}}$. Implementations are available for $N_1=2$, $N_2\in\{1,2,\infty\}$.